finished delete button

+++ app/src/routes.py
@@ -79,23 +79,26 @@ def logout():
 def account():
     logo_form = LogoUploadForm()
     delete_form = DeleteAccountForm()
+    print(delete_form.confirm.data)
 
-    if logo_form.validate_on_submit():
-        if logo_form.logo.data:
-            logo_filename = secure_filename(logo_form.logo.data.filename)
-            logo_path = os.path.join(app.root_path, "static", "logos", logo_filename)
-            logo_form.logo.data.save(logo_path)
-            current_user.logo_url = f"logos/{logo_filename}"
-            db.session.commit()
-            flash("Logo uploaded successfully!", "success")
+    if logo_form.validate_on_submit() and logo_form.logo.data:
+        logo_filename = secure_filename(logo_form.logo.data.filename)
+        logo_path = os.path.join(app.root_path, "static", "logos", logo_filename)
+        logo_form.logo.data.save(logo_path)
+        current_user.logo_url = f"logos/{logo_filename}"
+        db.session.commit()
+        flash("Logo uploaded successfully!", "success")
 
-    if delete_form.validate_on_submit():
-        if request.method == "POST":
-            db.session.delete(current_user)  # Delete user from the database
-            db.session.commit()
-            logout_user()  # Log the user out
-            flash("Your account has been deleted.", "success")
-            return redirect(url_for("home"))
+    elif (
+        delete_form.confirm.data
+        and delete_form.validate_on_submit()
+        and request.method == "POST"
+    ):
+        db.session.delete(current_user)
+        db.session.commit()
+        logout_user()
+        flash("Your account has been deleted.", "success")
+        return redirect(url_for("home"))
 
     return render_template(
         "account.html", title="Account", logo_form=logo_form, delete_form=delete_form
